Replicate a list of events into a data set.

Replicate a list of events into a data set.

assign_ev(l, idata, evgroup, join = FALSE)

Arguments

l

list of event objects

idata

an idata set (one ID per row)

evgroup

the character name of the column in idata that specifies event object to implement

join

if TRUE, join idata to the data set before returning.

Examples

ev1 <- ev(amt=100)
ev2 <- ev(amt=300, rate=100, ii=12, addl=10)

idata <- data.frame(ID=1:10)
idata$arm <- 1+(idata$ID %%2)

assign_ev(list(ev1,ev2),idata,"arm",join=TRUE)

#>    time cmt amt evid rate ii addl ID arm
#> 1     0   1 100    1    0  0    0  1   2
#> 2     0   1 300    1  100 12   10  2   1
#> 3     0   1 100    1    0  0    0  3   2
#> 4     0   1 300    1  100 12   10  4   1
#> 5     0   1 100    1    0  0    0  5   2
#> 6     0   1 300    1  100 12   10  6   1
#> 7     0   1 100    1    0  0    0  7   2
#> 8     0   1 300    1  100 12   10  8   1
#> 9     0   1 100    1    0  0    0  9   2
#> 10    0   1 300    1  100 12   10 10   1