Functions for chaining commands together.

Use these functions with chaining commands togehter with the operator.

Details

Other functions that may be used in the chain of commands include: param, init, update, ev. or any other function that will take the output of the preceeding command as it's first argument.

Examples

mod <- mrgsolve:::house()

data(exidata)
data(exTheoph)

out <- mod %>% data_set(exTheoph) %>% mrgsim()
out <- mod %>% carry_out(evid) %>% ev(amt=100, cmt=1) %>% mrgsim()
out <- mod %>% Req(CP,RESP) %>% mrgsim()