Do an mrgsim run

Do an mrgsim run

do_mrgsim(x, data, idata = null_idata, carry.out = character(0),
  mtime = numeric(0), seed = as.integer(NA), Request = character(0),
  capture = NULL, obsonly = FALSE, obsaug = FALSE, tgrid = numeric(0),
  recsort = 1, deslist = list(), descol = character(0), filbak = TRUE,
  tad = FALSE, nocb = TRUE, skip_init_calc = FALSE, ...)

Arguments

x
a model object

data
a simulation data set

idata
individual-level data

carry.out
data items to copy into the output

mtime
not used

seed
deprecated

Request
compartments or captured variables to retain in the simulated output

capture
character file name used for debugging (not related to $CAPTURE

obsonly
if TRUE, dosing records are not included in the output

obsaug
augment the data set with time grid observations

tgrid
a tgrid object

recsort
record sorting flag

deslist
a list of tgrid objects

descol
the name of a column for assigning designs

filbak
carry data items backward when the first data set row has time greater than zero

tad
if TRUE, include time after dose in the output

nocb
if TRUE, use next observation carry backward method; otherwise, use locf

skip_init_calc
don't use $MAIN to calculate initial conditions

...
not used