Schedule dosing events on days of the week.

This function lets you schedule doses on specific days of the week, allowing you to create dosing regimens on Monday/Wednesday/Friday, or Tuesday/Thursday, or every other day (however you want to define that) etc.

ev_days(ev = NULL, days = "", addl = 0, ii = 168, unit = c("hours",
  "days"), ...)

Arguments

ev

an event object

days

comma- or space-separated character string of valid days of the the week (see details)

addl

additional doses to administer

ii

inter-dose interval; intended use is to keep this at the default value

unit

time unit; the function can only currently handle hours or days

...

event objects named by one the valid days of the week (see details)

Details

Valid names of the week are:

m for Monday

t for Tuesday

w for Wednesday

th for Thursday

f for Friday

sa for Saturday

s for Sunday

The whole purpose of this function is to schedule doses on specific days of the week, in a repeating weekly schedule. Please do use caution when changing ii from it's default value.

Examples

# Monday, Wednesday, Friday x 4 weeks
ev_days(ev(amt=100), days="m,w,f", addl=3)

# 50 mg Tuesdays, 100 mg Thursdays x 6 months
ev_days(t=ev(amt=50), th=ev(amt=100), addl=23)