Schedule a series of event objects

Schedule a series of event objects

Usage

ev_seq(..., ID = NULL, .dots = NULL, id = NULL)

# S3 method for ev
seq(...)

Arguments

...

event objects or numeric arguments named wait

ID

numeric vector of subject IDs

.dots

a list of event objects that replaces ...

id

deprecated; use ID

Value

A single event object.

Details

The doses for the next event line start after all of the doses from the previous event line plus one dosing interval from the previous event line (see examples).

When numerics named wait are mixed in with the event objects, a period with no dosing activity is incorporated into the sequence, between the adjacent dosing event objects. Values for wait can be negative.

Values for time in any event object act like a prefix time spacer wherever that event occurs in the event sequence (see examples).

Use the generic seq when the first argument is an event object. If a waiting period is the first event, you will need to use ev_seq. When an event object has multiple rows, the end time for that sequence is taken to be one dosing interval after the event that takes place on the last row of the event object.

Examples

e1 <- ev(amt=100, ii=12, addl=1)

e2 <- ev(amt=200)

seq(e1, e2)
#> Events:
#>   time amt ii addl cmt evid
#> 1    0 100 12    1   1    1
#> 2   24 200  0    0   1    1

seq(e1, .ii = 8, e2)
#> Events:
#>   time amt ii addl cmt evid
#> 1    0 100 12    1   1    1
#> 2   20 200  0    0   1    1

seq(e1, wait = 8, e2)
#> Events:
#>   time amt ii addl cmt evid
#> 1    0 100 12    1   1    1
#> 2   32 200  0    0   1    1

seq(e1, .ii = 8, e2, ID = 1:10)
#> Events:
#>    ID time amt ii addl cmt evid
#> 1   1    0 100 12    1   1    1
#> 2   1   20 200  0    0   1    1
#> 3   2    0 100 12    1   1    1
#> 4   2   20 200  0    0   1    1
#> 5   3    0 100 12    1   1    1
#> 6   3   20 200  0    0   1    1
#> 7   4    0 100 12    1   1    1
#> 8   4   20 200  0    0   1    1
#> 9   5    0 100 12    1   1    1
#> 10  5   20 200  0    0   1    1
#> 11  6    0 100 12    1   1    1
#> 12  6   20 200  0    0   1    1
#> 13  7    0 100 12    1   1    1
#> 14  7   20 200  0    0   1    1
#> 15  8    0 100 12    1   1    1
#> 16  8   20 200  0    0   1    1
#> 17  9    0 100 12    1   1    1
#> 18  9   20 200  0    0   1    1
#> 19 10    0 100 12    1   1    1
#> 20 10   20 200  0    0   1    1

ev_seq(.ii = 12, e1, .ii = 120, e2, .ii = 120, e1)
#> Events:
#>   time amt ii addl cmt evid
#> 1   12 100 12    1   1    1
#> 2  144 200  0    0   1    1
#> 3  264 100 12    1   1    1

seq(ev(amt=100, ii=12), ev(time=8, amt=200))
#> Events:
#>   time amt ii addl cmt evid
#> 1    0 100 12    0   1    1
#> 2   20 200  0    0   1    1