Create an event object with data-like names

This function calls ev() to create an event object and then sets the case attribute so that it renders nmtran data names in upper case. An object created with evd() can be used in the same way as an object created with ev().

evd(x, ...)

# S4 method for mrgmod
evd(x, ...)

# S4 method for missing
evd(x, ...)

# S4 method for ev
evd(x, ...)

as.evd(x)

Arguments

x

An event object.

...

Arguments passed to ev().

Details

Note that evd isn't a separate class; it is just an ev object with a specific case attribute. See examples which illustrate the difference.

See also

Examples

a <- evd(amt = 100)
b <- ev(amt = 300)
a
#> Events Data:
#>   time amt cmt evid
#> 1    0 100   1    1
as.data.frame(a)
#>   TIME AMT CMT EVID
#> 1    0 100   1    1
as_data_set(a, b)
#>   ID TIME CMT EVID AMT
#> 1  1    0   1    1 100
#> 2  2    0   1    1 300
as_data_set(b, a)
#>   ID time cmt evid amt
#> 1  1    0   1    1 300
#> 2  2    0   1    1 100
as.data.frame(seq(a, b))
#>   TIME AMT II ADDL CMT EVID
#> 1    0 100  0    0   1    1
#> 2    0 300  0    0   1    1