Check whether all required parameters needed in a model are present in an object

Check whether all required parameters needed in a model are present in an object

inventory(x, obj, ..., .strict = FALSE)

Arguments

x
model object

obj
data.frame to pass to idata_set or data_set

...
capture dplyr-style parameter requirements

.strict
whether to stop execution if all requirements are present (TRUE) or just warn (FALSE); see details

Value

original mrgmod

Details

If parameter requirements are not explicitly stated, the requirement defaults to all parameter names in x. Note that, by default, the inventory is not .strict unless the user explicitly states the parameter requirement. That is, if parameter requirements are explicitly stated, .strict will be set to TRUE if a value .strict was not passed in the call.

Examples

# NOT RUN {
inventory(mod, idata, CL:V) # parameters defined, inclusively, CL through Volume 
inventory(mod, idata, everything()) # all parameters
inventory(mod, idata, contains("OCC")) # all parameters containing OCC
inventory(mod, idata, -F) # all parameters except F
# }