Run sensitivity analysis on model settings.

Knobs can be parameter values or PK dosing items (e.g. amt). By design, all combinations of specified knob/values are simulated.

knobs(x, y, ...)

# S4 method for mrgmod,missing
knobs(x, y, ...)

# S4 method for mrgmod,batch_mrgsims
knobs(x, y, ...)

# S4 method for batch_mrgsims
as.data.frame(x, row.names = NULL,
  optional = FALSE, ...)

# S4 method for batch_mrgsims,ANY
knobs(x, y, ...)

# S4 method for batch_mrgsims
show(object)

Arguments

x
the model object

y
a batch_mrgsims object

...
knobs: named numeric vectors that identify knob names and knob values for a batch run. See details.

row.names
passed to as.data.frame

optional
passed to as.data.frame

object
passed to show

Value

An object of class batch_mrgsims. Most methods for mrgsims objects also work on batch_mrgsims object.

Details

Valid knob names include: any parameter name (in param(mod)), time variables (start, end, delta), PK dosing items (amt, ii, rate, and others ...), and solver settings (atol, hmax, etc...).

Examples

## example("knobs")

mod <- mrgsolve:::house(end=72)

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

events <- ev(amt=1000, cmt=1, addl=3, ii=12)

out <- mod %>% ev(events) %>% knobs(CL=c(1,2,3))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out)

#> Error in plot(out): object 'out' not found

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- mod %>% ev(events) %>% knobs(CL=c(1,2,3), VC=c(5,20,50))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out)

#> Error in plot(out): object 'out' not found

plot(out,CP~.)

#> Error in plot(out, CP ~ .): object 'out' not found

plot(out, CP~time|VC, groups=CL, lty=2)

#> Error in plot(out, CP ~ time | VC, groups = CL, lty = 2): object 'out' not found

out <- knobs(mod, amt=c(100,300,500), cmt=1)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), structure("missing", package = "methods")),     structure(function (x, y, ...)     standardGeneric("knobs"), generic = structure("knobs", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "y"), default = `\001NULL\001`, skeleton = (function (x,         y, ...)     stop("invalid call in method dispatch to 'knobs' (no default method)",         domain = NA))(x, y, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘knobs’ for signature ‘"standardGeneric", "missing"’

plot(out)

#> Error in plot(out): object 'out' not found

out <- mod %>% knobs(amt=c(100,300), CL=c(1,3), VC=c(5,20), cmt=1)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), structure("missing", package = "methods")),     structure(function (x, y, ...)     standardGeneric("knobs"), generic = structure("knobs", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "y"), default = `\001NULL\001`, skeleton = (function (x,         y, ...)     stop("invalid call in method dispatch to 'knobs' (no default method)",         domain = NA))(x, y, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘knobs’ for signature ‘"standardGeneric", "missing"’

plot(out)

#> Error in plot(out): object 'out' not found

plot(out, CP~.)

#> Error in plot(out, CP ~ .): object 'out' not found

out <- knobs(mod, CL=c(1,2,3))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), structure("missing", package = "methods")),     structure(function (x, y, ...)     standardGeneric("knobs"), generic = structure("knobs", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "y"), default = `\001NULL\001`, skeleton = (function (x,         y, ...)     stop("invalid call in method dispatch to 'knobs' (no default method)",         domain = NA))(x, y, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘knobs’ for signature ‘"standardGeneric", "missing"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- knobs(mod, CL=c(1,2,3))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), structure("missing", package = "methods")),     structure(function (x, y, ...)     standardGeneric("knobs"), generic = structure("knobs", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "y"), default = `\001NULL\001`, skeleton = (function (x,         y, ...)     stop("invalid call in method dispatch to 'knobs' (no default method)",         domain = NA))(x, y, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘knobs’ for signature ‘"standardGeneric", "missing"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found