Write, compile, and load model code.

This is a convenience function that ultimately calls mread.

mcode_cache(model, code, project = tempdir(), ...)

mcode(model, code, project = tempdir(), ...)

Arguments

model

model name

code

character string specifying a mrgsolve model

project

project name

...

passed to mread

Details

Note that the arguments are in slightly different order than mread. The default project is tempdir().

Examples

code <- '
$CMT DEPOT CENT
$PKMODEL ncmt=1, depot=TRUE
$MAIN
double CL = 1;
double V = 20;
double KA = 1;
'

mod <- mcode("example",code)

#> Warning: path[1]="C:\Users\devin\Documents\Repos\mrgsolve/base": The system cannot find the file specified

#> Compiling example ... 

#> c:/Rtools/mingw_64/bin/g++  -I"C:/PROGRA~1/R/R-33~1.2/include" -DNDEBUG    -I"C:/Users/devin/Documents/Repos/mrgsolve/base" -I"C:/Users/devin/AppData/Local/Temp/RtmpCUybu4" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c example-mread-source.cpp -o example-mread-source.o
#> example-mread-source.cpp:13:25: fatal error: modelheader.h: No such file or directory
#>  #include "modelheader.h"
#>                          ^
#> compilation terminated.
#> make: *** [example-mread-source.o] Error 1
#> Warning message:
#> running command 'make -f "C:/PROGRA~1/R/R-33~1.2/etc/x64/Makeconf" -f "C:/PROGRA~1/R/R-33~1.2/share/make/winshlib.mk" SHLIB_LDFLAGS='$(SHLIB_CXXLDFLAGS)' SHLIB_LD='$(SHLIB_CXXLD)' SHLIB="example-mread-source.dll" WIN=64 TCLBIN=64 OBJECTS="example-mread-source.o"' had status 2 
#> 
#> 

#> Error: There was a problem when compiling the model. c:/Rtools/mingw_64/bin/g++  -I"C:/PROGRA~1/R/R-33~1.2/include" -DNDEBUG    -I"C:/Users/devin/Documents/Repos/mrgsolve/base" -I"C:/Users/devin/AppData/Local/Temp/RtmpCUybu4" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c example-mread-source.cpp -o example-mread-source.oThere was a problem when compiling the model. example-mread-source.cpp:13:25: fatal error: modelheader.h: No such file or directoryThere was a problem when compiling the model.  #include "modelheader.h"There was a problem when compiling the model.                          ^There was a problem when compiling the model. compilation terminated.There was a problem when compiling the model. make: *** [example-mread-source.o] Error 1There was a problem when compiling the model. Warning message:There was a problem when compiling the model. running command 'make -f "C:/PROGRA~1/R/R-33~1.2/etc/x64/Makeconf" -f "C:/PROGRA~1/R/R-33~1.2/share/make/winshlib.mk" SHLIB_LDFLAGS='$(SHLIB_CXXLDFLAGS)' SHLIB_LD='$(SHLIB_CXXLD)' SHLIB="example-mread-source.dll" WIN=64 TCLBIN=64 OBJECTS="example-mread-source.o"' had status 2