modlib: PK/PD Model parameters, compartments, and output variables.

modlib: PK/PD Model parameters, compartments, and output variables.

Compartments

EV1, EV2: extravasular dosing compartments

CENT: central PK compartment

PERIPH: peripheral PK compartment

PERIPH2: peripheral PK compartment 2

RESP: response PD compartment (irm models)

Output variables

CP: concentration in the central compartment (CENT/VC)

RESP: response (emax model)

PK parameters

KA1, KA2: first order absorption rate constants from first and second extravascular compartment (1/time)

CL: clearance (volume/time)

VC: volume of distribution, central compartment (volume)

VP: volume of distribution, peripheral compartment (volume)

VP2: volume of distribution, peripheral compartment 2 (volume)

Q: intercompartmental clearance (volume/time)

Q2: intercompartmental clearance 2 (volume/time)

VMAX: maximum rate, nonlinear process (mass/time)

KM: Michaelis constant (mass/volume)

K10: elimination rate constant (1/time); CL/VC

K12: rate constant for transfer to peripheral compartment from central (1/time); Q/VC

K21: rate constant for transfer to central compartment from peripheral (1/time); Q/VP

PD parameters

E0: baseline effect (emax model)

EMAX, IMAX: maximum effect (response)

EC50, IC50: concentration producing 50 percent of effect (mass/volume)

KIN: zero-order response production rate (irm models) (response/time)

KOUT: first-order response elimination rate (irm models) (1/time)

n: sigmoidicity factor

KEO: rate constant for transfer to effect compartment (1/time)