Sensitivity Analysis with 'mrgsolve'

Perform local sensitivity analysis on ordinary differential equation based models, including ad-hoc graphical analyses based on sequences of parameters as well as local sensitivity analysis. Functions are provided for creating inputs, simulating scenarios and plotting outputs.

Details

Local sensitivity analysis: lsa(), lsa_plot()

Run ad-hoc sensitivity analyses: sens_each(), sens_grid(), sens_run()

Use sens_each_data() and sens_grid_data() to pass in data sets

Parameter sequence generation:

In a pipeline: parseq_cv(), parseq_fct(), parseq_range(), parseq_manual()

Stand alone: seq_cv(), seq_fct(), seq_geo(), seq_even()

Plot ad-hoc sensitivity analysis results

Use sens_plot()

Select parameters or results

Use select_par() to choose which parameters to vary

Use select_sens() to subset sensitivity analysis results

See also