Methods for working with

These methods help the user view simulation output and extract simulated data to work with further. The methods listed here for the most part have generics defined by R or other R packages. See the seealso section for other methods defined by mrgsolve that have their own documentation pages.

# S4 method for mrgsims
$(x, name)

# S4 method for mrgsims
tail(x, ...)

# S4 method for mrgsims
head(x, ...)

# S4 method for mrgsims
dim(x)

# S4 method for mrgsims
names(x)

# S4 method for mrgsims
as.data.frame(x, row.names = NULL, optional = FALSE,
  ...)

# S4 method for mrgsims
as.matrix(x, ...)

# S4 method for mrgsims
subset(x, ...)

# S4 method for mrgsims
summary(object, ...)

# S4 method for mrgsims
show(object)

Arguments

x

mrgsims object

name

name of column of simulated output to retain

...

passed to other functions

row.names

passed to as.data.frame

optional

passed to as.data.frame

object

passed to show

Details

Most methods should behave as expected according to other method commonly used in R (e.g. head, tail, as.data.frame, etc ...)

subset coreces simulated output to data.frame and passes to subset.data.frame

$ selects a column in the simulated data and returns numeric

head see head.matrix; returns simulated data

tail see tail.matrix; returns simulated data

dim, nrow, ncol returns dimensions, number of rows, and number of columns in simulated data

as.data.frame coreces simulated data to data.frame and returns the data.frame

as.matrix returns matrix of simulated data

as.tbl coreces simulated to tbl_df; requires dplyr

summary coreces simulated data to data.frame and passes to summary.data.frame

plot plots simulated data; see plot_mrgsims

See also

stime

Examples

## example("mrgsims")

mod <- mrgsolve:::house() %>% init(GUT=100)

out <- mrgsim(mod)
class(out)

out
head(out)
tail(out)

mrgsolve:::mod(out)

dim(out)
names(out)

mat <- as.matrix(out)
df <- as.data.frame(out)

df <- subset(out, time < 12) ## a data frame
out$CP

plot(out)
plot(out, CP~.)
plot(out, CP+RESP~time, scales="same", xlab="Time", main="Model sims")