Get THETA, OMEGA and SIGMA from a completed NONMEM run

Get THETA, OMEGA and SIGMA from a completed NONMEM run

nmxml(run = numeric(0), project = character(0), file = character(0),
  theta = TRUE, omega = FALSE, sigma = FALSE, olabels = NULL,
  slabels = NULL, oprefix = "", sprefix = "", tname = "THETA",
  oname = "...", sname = "...", ...)

Arguments

run

run number

project

project directory

file

the complete path to the run.xml file

theta

logical; if TRUE, the $THETA vector is returned

omega

logical; if TRUE, the $OMEGA matrix is returned

sigma

logical; if TRUE, the $SIGMA matrix is returned

olabels

labels for $OMEGA

slabels

labels for $SIGMA

oprefix

prefix for $OMEGA labels

sprefix

prefix for $SIGMA labels

tname

name for $THETA

oname

name for $OMEGA

sname

name for $SIGMA

...

passed along

Value

a list with theta, omega and sigma elements, depending on what was requested

Details

If run and project are supplied, the .xml file is assumed to be located in run.xml, in directory run off the project directory. If file is supplied, run and project arguments are ignored.