Create and work with parameter objects.

See numericlist for methods to deal with parameter_list objects.

# S4 method for mrgmod
$(x, name)

param(.x, ...)

# S4 method for mrgmod
param(.x, .y = list(), ..., .pat = "*",
  .strict = FALSE)

# S4 method for mrgsims
param(.x, ...)

# S4 method for missing
param(..., .strict = TRUE)

# S4 method for list
param(.x, ...)

# S4 method for ANY
param(.x, ...)

as.param(.x, ...)

# S4 method for list
as.param(.x, ...)

# S4 method for numeric
as.param(.x, ...)

# S4 method for parameter_list
as.param(.x, ...)

# S4 method for missing
as.param(.x, ...)

# S4 method for parameter_list
show(object)

allparam(.x)

Arguments

x

mrgmod object

name

parameter to take

.x

the model object

...

passed along or name/value pairs to update the parameters in a model object

.y

list to be merged into parameter list

.pat

a regular expression (character) to be applied as a filter for which parameters to show when printing

.strict

if TRUE, all names to be updated must be found in the parameter list

object

passed to show

Value

An object of class parameter_list (see numericlist).

Details

Can be used to either get a parameter list object from a mrgmod model object or to update the parameters in a model object. For both uses, the return value is a parameter_list object. For the former use, param is usually called to print the parameters to the screen, but the parameter_list object can also be coreced to a list or numeric R object.

Examples

## example("param")
mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve.git/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

param(mod)

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers

param(mod, .pat="^(C|F)") ## may be useful when large number of parameters

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers

class(param(mod))

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers

param(mod)$KA

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers

as.list(param(mod))

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers

as.data.frame(param(mod))

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: Invalid names
#> invalid class “parameter_list” object: 2: All parameters require names
#> invalid class “parameter_list” object: 3: All parameters must be single numbers