Simulate from 1- or 2-compartment PK model.

This is an R function that returns model objects based on $PKMODEL.

pkmodel(ncmt = 1, depot = FALSE, ...)

Arguments

ncmt

passed to PKMODEL

depot

passed to PKMODEL

...

passed to update

Value

An object of class mrgmod-class

Details

Once the model object is generated, use param to check names of the parameters in the model and init to check the names of the compartments in the model. Calculations for the amounts in each compartment are done via analytical solutions, not differential equations. A subject-level random effect is also provided for each PK parameter; use omat to see the names of those random effects. All random effect variances have initial value of zero and may be updated via omat.

Examples

mod <- pkmodel(1)

#> Warning: path[1]="C:\Users\devin\Documents\Repos\mrgsolve/base": The system cannot find the file specified

#> c:/Rtools/mingw_64/bin/g++  -I"C:/PROGRA~1/R/R-33~1.2/include" -DNDEBUG    -I"C:/Users/devin/Documents/Repos/mrgsolve/base" -I"C:/Users/devin/AppData/Local/Temp/RtmpCUybu4" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c modpk1-mread-source.cpp -o modpk1-mread-source.o
#> modpk1-mread-source.cpp:13:25: fatal error: modelheader.h: No such file or directory
#>  #include "modelheader.h"
#>                          ^
#> compilation terminated.
#> make: *** [modpk1-mread-source.o] Error 1
#> Warning message:
#> running command 'make -f "C:/PROGRA~1/R/R-33~1.2/etc/x64/Makeconf" -f "C:/PROGRA~1/R/R-33~1.2/share/make/winshlib.mk" SHLIB_LDFLAGS='$(SHLIB_CXXLDFLAGS)' SHLIB_LD='$(SHLIB_CXXLD)' SHLIB="modpk1-mread-source.dll" WIN=64 TCLBIN=64 OBJECTS="modpk1-mread-source.o"' had status 2 
#> 
#> 

#> Error: There was a problem when compiling the model. c:/Rtools/mingw_64/bin/g++  -I"C:/PROGRA~1/R/R-33~1.2/include" -DNDEBUG    -I"C:/Users/devin/Documents/Repos/mrgsolve/base" -I"C:/Users/devin/AppData/Local/Temp/RtmpCUybu4" -I"d:/Compiler/gcc-4.9.3/local330/include"     -O2 -Wall  -mtune=core2 -c modpk1-mread-source.cpp -o modpk1-mread-source.oThere was a problem when compiling the model. modpk1-mread-source.cpp:13:25: fatal error: modelheader.h: No such file or directoryThere was a problem when compiling the model.  #include "modelheader.h"There was a problem when compiling the model.                          ^There was a problem when compiling the model. compilation terminated.There was a problem when compiling the model. make: *** [modpk1-mread-source.o] Error 1There was a problem when compiling the model. Warning message:There was a problem when compiling the model. running command 'make -f "C:/PROGRA~1/R/R-33~1.2/etc/x64/Makeconf" -f "C:/PROGRA~1/R/R-33~1.2/share/make/winshlib.mk" SHLIB_LDFLAGS='$(SHLIB_CXXLDFLAGS)' SHLIB_LD='$(SHLIB_CXXLD)' SHLIB="modpk1-mread-source.dll" WIN=64 TCLBIN=64 OBJECTS="modpk1-mread-source.o"' had status 2 

mod %>% ev(amt=1000, ii=24, addl=3) %>% mrgsim(end=120)

#> Model:  demo.cpp 
#> Dim:    1202 x 6 
#> Time:   0 to 120 
#> ID:     1 
#>      ID time    GUT  CENT  RESP    CP
#> [1,]  1  0.0    0.0   0.0 50.00  0.00
#> [2,]  1  0.0 1000.0   0.0 50.00  0.00
#> [3,]  1  0.1  878.1 121.6 47.92  6.08
#> [4,]  1  0.2  771.1 227.8 44.07 11.39
#> [5,]  1  0.3  677.1 320.4 39.90 16.02
#> [6,]  1  0.4  594.5 401.1 35.89 20.06
#> [7,]  1  0.5  522.0 471.4 32.23 23.57
#> [8,]  1  0.6  458.4 532.5 28.96 26.63

## Not run: ------------------------------------
# mod <- pkmodel(1,TRUE)
# mod <- pkmodel(2)
# mod <- pkmodel(2,TRUE)
## ---------------------------------------------