Plot a lsa object

Plot a lsa object

Usage

# S3 method for class 'lsa'
plot(x, y = NULL, palette = NULL, pal = NULL, ...)

Arguments

x

output from lsa().

y

not used.

palette

a discrete color scale object, such as one returned by ggplot2::scale_color_brewer() or ggplot2::scale_color_manual(). When NULL (default), a palette is chosen automatically based on the number of parameters.

pal

; please use palette instead.

...

not used.

Value

A ggplot.