Get the times at which the model will be evaluated.

Get the times at which the model will be evaluated.

stime(x, ...)

Arguments

x

object of class mrgmod

...

passed on

Value

a sorted vector of unique times

Details

Simulation times include the sequence of times created from start, end, and delta and the vector of times found in add. Making end negative will omit any start / end / delta sequence. Negative values are discarded from the result.

Examples

## example("stime", package="mrgsolve")

mod <- mrgsolve:::house(end=12, delta=2, add=c(11,13,15))

stime(mod)