Validate and prepare a data sets for simulation

This function is called by mrgsim. Users may also call this function to pre-validate data when the same data set is used for repeated simulation.

valid_data_set(x, m = NULL, verbose = FALSE, quiet = FALSE)

valid_data_set.matrix(x, verbose = FALSE)

Arguments

x
data.frame or matrix

m
a model object

verbose
logical

quiet
if TRUE, messages will be suppressed

Value

A matrix with non-numeric columns dropped; if x is a data.frame with character cmt column comprised of valid compartment names and m is a model object, the cmt column will be converted to the corresponding compartment number.

See also

Examples

mod <- mrgsolve:::house()

data(exTheoph)

valid_data_set(exTheoph,mod)

#>     ID   WT Dose  time  conc cmt  amt evid ..zeros..
#> 1    1 79.6 4.02  0.00  0.00   1 4.02    1         0
#> 2    1 79.6 4.02  0.25  2.84   0 0.00    0         0
#> 3    1 79.6 4.02  0.57  6.57   0 0.00    0         0
#> 4    1 79.6 4.02  1.12 10.50   0 0.00    0         0
#> 5    1 79.6 4.02  2.02  9.66   0 0.00    0         0
#> 6    1 79.6 4.02  3.82  8.58   0 0.00    0         0
#> 7    1 79.6 4.02  5.10  8.36   0 0.00    0         0
#> 8    1 79.6 4.02  7.03  7.47   0 0.00    0         0
#> 9    1 79.6 4.02  9.05  6.89   0 0.00    0         0
#> 10   1 79.6 4.02 12.12  5.94   0 0.00    0         0
#> 11   1 79.6 4.02 24.37  3.28   0 0.00    0         0
#> 12   2 72.4 4.40  0.00  0.00   1 4.40    1         0
#> 13   2 72.4 4.40  0.27  1.72   0 0.00    0         0
#> 14   2 72.4 4.40  0.52  7.91   0 0.00    0         0
#> 15   2 72.4 4.40  1.00  8.31   0 0.00    0         0
#> 16   2 72.4 4.40  1.92  8.33   0 0.00    0         0
#> 17   2 72.4 4.40  3.50  6.85   0 0.00    0         0
#> 18   2 72.4 4.40  5.02  6.08   0 0.00    0         0
#> 19   2 72.4 4.40  7.03  5.40   0 0.00    0         0
#> 20   2 72.4 4.40  9.00  4.55   0 0.00    0         0
#> 21   2 72.4 4.40 12.00  3.01   0 0.00    0         0
#> 22   2 72.4 4.40 24.30  0.90   0 0.00    0         0
#> 23   3 70.5 4.53  0.00  0.00   1 4.53    1         0
#> 24   3 70.5 4.53  0.27  4.40   0 0.00    0         0
#> 25   3 70.5 4.53  0.58  6.90   0 0.00    0         0
#> 26   3 70.5 4.53  1.02  8.20   0 0.00    0         0
#> 27   3 70.5 4.53  2.02  7.80   0 0.00    0         0
#> 28   3 70.5 4.53  3.62  7.50   0 0.00    0         0
#> 29   3 70.5 4.53  5.08  6.20   0 0.00    0         0
#> 30   3 70.5 4.53  7.07  5.30   0 0.00    0         0
#> 31   3 70.5 4.53  9.00  4.90   0 0.00    0         0
#> 32   3 70.5 4.53 12.15  3.70   0 0.00    0         0
#> 33   3 70.5 4.53 24.17  1.05   0 0.00    0         0
#> 34   4 72.7 4.40  0.00  0.00   1 4.40    1         0
#> 35   4 72.7 4.40  0.35  1.89   0 0.00    0         0
#> 36   4 72.7 4.40  0.60  4.60   0 0.00    0         0
#> 37   4 72.7 4.40  1.07  8.60   0 0.00    0         0
#> 38   4 72.7 4.40  2.13  8.38   0 0.00    0         0
#> 39   4 72.7 4.40  3.50  7.54   0 0.00    0         0
#> 40   4 72.7 4.40  5.02  6.88   0 0.00    0         0
#> 41   4 72.7 4.40  7.02  5.78   0 0.00    0         0
#> 42   4 72.7 4.40  9.02  5.33   0 0.00    0         0
#> 43   4 72.7 4.40 11.98  4.19   0 0.00    0         0
#> 44   4 72.7 4.40 24.65  1.15   0 0.00    0         0
#> 45   5 54.6 5.86  0.00  0.00   1 5.86    1         0
#> 46   5 54.6 5.86  0.30  2.02   0 0.00    0         0
#> 47   5 54.6 5.86  0.52  5.63   0 0.00    0         0
#> 48   5 54.6 5.86  1.00 11.40   0 0.00    0         0
#> 49   5 54.6 5.86  2.02  9.33   0 0.00    0         0
#> 50   5 54.6 5.86  3.50  8.74   0 0.00    0         0
#> 51   5 54.6 5.86  5.02  7.56   0 0.00    0         0
#> 52   5 54.6 5.86  7.02  7.09   0 0.00    0         0
#> 53   5 54.6 5.86  9.10  5.90   0 0.00    0         0
#> 54   5 54.6 5.86 12.00  4.37   0 0.00    0         0
#> 55   5 54.6 5.86 24.35  1.57   0 0.00    0         0
#> 56   6 80.0 4.00  0.00  0.00   1 4.00    1         0
#> 57   6 80.0 4.00  0.27  1.29   0 0.00    0         0
#> 58   6 80.0 4.00  0.58  3.08   0 0.00    0         0
#> 59   6 80.0 4.00  1.15  6.44   0 0.00    0         0
#> 60   6 80.0 4.00  2.03  6.32   0 0.00    0         0
#> 61   6 80.0 4.00  3.57  5.53   0 0.00    0         0
#> 62   6 80.0 4.00  5.00  4.94   0 0.00    0         0
#> 63   6 80.0 4.00  7.00  4.02   0 0.00    0         0
#> 64   6 80.0 4.00  9.22  3.46   0 0.00    0         0
#> 65   6 80.0 4.00 12.10  2.78   0 0.00    0         0
#> 66   6 80.0 4.00 23.85  0.92   0 0.00    0         0
#> 67   7 64.6 4.95  0.00  0.00   1 4.95    1         0
#> 68   7 64.6 4.95  0.25  0.85   0 0.00    0         0
#> 69   7 64.6 4.95  0.50  2.35   0 0.00    0         0
#> 70   7 64.6 4.95  1.02  5.02   0 0.00    0         0
#> 71   7 64.6 4.95  2.02  6.58   0 0.00    0         0
#> 72   7 64.6 4.95  3.48  7.09   0 0.00    0         0
#> 73   7 64.6 4.95  5.00  6.66   0 0.00    0         0
#> 74   7 64.6 4.95  6.98  5.25   0 0.00    0         0
#> 75   7 64.6 4.95  9.00  4.39   0 0.00    0         0
#> 76   7 64.6 4.95 12.05  3.53   0 0.00    0         0
#> 77   7 64.6 4.95 24.22  1.15   0 0.00    0         0
#> 78   8 70.5 4.53  0.00  0.00   1 4.53    1         0
#> 79   8 70.5 4.53  0.25  3.05   0 0.00    0         0
#> 80   8 70.5 4.53  0.52  3.05   0 0.00    0         0
#> 81   8 70.5 4.53  0.98  7.31   0 0.00    0         0
#> 82   8 70.5 4.53  2.02  7.56   0 0.00    0         0
#> 83   8 70.5 4.53  3.53  6.59   0 0.00    0         0
#> 84   8 70.5 4.53  5.05  5.88   0 0.00    0         0
#> 85   8 70.5 4.53  7.15  4.73   0 0.00    0         0
#> 86   8 70.5 4.53  9.07  4.57   0 0.00    0         0
#> 87   8 70.5 4.53 12.10  3.00   0 0.00    0         0
#> 88   8 70.5 4.53 24.12  1.25   0 0.00    0         0
#> 89   9 86.4 3.10  0.00  0.00   1 3.10    1         0
#> 90   9 86.4 3.10  0.30  7.37   0 0.00    0         0
#> 91   9 86.4 3.10  0.63  9.03   0 0.00    0         0
#> 92   9 86.4 3.10  1.05  7.14   0 0.00    0         0
#> 93   9 86.4 3.10  2.02  6.33   0 0.00    0         0
#> 94   9 86.4 3.10  3.53  5.66   0 0.00    0         0
#> 95   9 86.4 3.10  5.02  5.67   0 0.00    0         0
#> 96   9 86.4 3.10  7.17  4.24   0 0.00    0         0
#> 97   9 86.4 3.10  8.80  4.11   0 0.00    0         0
#> 98   9 86.4 3.10 11.60  3.16   0 0.00    0         0
#> 99   9 86.4 3.10 24.43  1.12   0 0.00    0         0
#> 100 10 58.2 5.50  0.00  0.00   1 5.50    1         0
#> 101 10 58.2 5.50  0.37  2.89   0 0.00    0         0
#> 102 10 58.2 5.50  0.77  5.22   0 0.00    0         0
#> 103 10 58.2 5.50  1.02  6.41   0 0.00    0         0
#> 104 10 58.2 5.50  2.05  7.83   0 0.00    0         0
#> 105 10 58.2 5.50  3.55 10.21   0 0.00    0         0
#> 106 10 58.2 5.50  5.05  9.18   0 0.00    0         0
#> 107 10 58.2 5.50  7.08  8.02   0 0.00    0         0
#> 108 10 58.2 5.50  9.38  7.14   0 0.00    0         0
#> 109 10 58.2 5.50 12.10  5.68   0 0.00    0         0
#> 110 10 58.2 5.50 23.70  2.42   0 0.00    0         0
#> 111 11 65.0 4.92  0.00  0.00   1 4.92    1         0
#> 112 11 65.0 4.92  0.25  4.86   0 0.00    0         0
#> 113 11 65.0 4.92  0.50  7.24   0 0.00    0         0
#> 114 11 65.0 4.92  0.98  8.00   0 0.00    0         0
#> 115 11 65.0 4.92  1.98  6.81   0 0.00    0         0
#> 116 11 65.0 4.92  3.60  5.87   0 0.00    0         0
#> 117 11 65.0 4.92  5.02  5.22   0 0.00    0         0
#> 118 11 65.0 4.92  7.03  4.45   0 0.00    0         0
#> 119 11 65.0 4.92  9.03  3.62   0 0.00    0         0
#> 120 11 65.0 4.92 12.12  2.69   0 0.00    0         0
#> 121 11 65.0 4.92 24.08  0.86   0 0.00    0         0
#> 122 12 60.5 5.30  0.00  0.00   1 5.30    1         0
#> 123 12 60.5 5.30  0.25  1.25   0 0.00    0         0
#> 124 12 60.5 5.30  0.50  3.96   0 0.00    0         0
#> 125 12 60.5 5.30  1.00  7.82   0 0.00    0         0
#> 126 12 60.5 5.30  2.00  9.72   0 0.00    0         0
#> 127 12 60.5 5.30  3.52  9.75   0 0.00    0         0
#> 128 12 60.5 5.30  5.07  8.57   0 0.00    0         0
#> 129 12 60.5 5.30  7.07  6.59   0 0.00    0         0
#> 130 12 60.5 5.30  9.03  6.11   0 0.00    0         0
#> 131 12 60.5 5.30 12.05  4.57   0 0.00    0         0
#> 132 12 60.5 5.30 24.15  1.17   0 0.00    0         0
#> attr(,"class")
#> [1] "valid_data_set"