Parse PKMODEL BLOCK data

Parse PKMODEL BLOCK data

PKMODEL(
  ncmt = 1,
  depot = FALSE,
  cmt = NULL,
  trans = pick_trans(ncmt, depot),
  env = list(),
  pos = 1,
  ...
)

Arguments

ncmt
number of compartments; must be 1 (one-compartment, not including a depot dosing compartment) or 2 (two-compartment model, not including a depot dosing compartment)

depot
logical indicating whether to add depot compartment

cmt
compartment names as comma-delimited character

trans
the parameterization for the PK model; must be 1, 2, 4, or 11

env
parse environment

pos
block position number

...
not used

Details

When using $PKMODEL, certain symbols must be defined in the model specification depending on the value of ncmt, depot and trans.

ncmt 1, depot FALSE, trans 2: CL, V

ncmt 1, depot TRUE , trans 2: CL, V, KA

ncmt 2, depot FALSE, trans 4: CL, V1, Q, V2

ncmt 2, depot TRUE , trans 4: CL, V2, Q, V3, KA

If trans=11 is specified, use the symbols listed above for the ncmt / depot combination, but append i at the end (e.g. CLi or Qi or KAi).

If trans=1, the user must utilize the following symbols:

pred_CL for clearance

pred_V or pred_V2 for central compartment volume of distribution

pred_Q for intercompartmental clearance

pred_V3 for for peripheral compartment volume of distribution

pred_KA for absorption rate constant

See also