Functions for chaining commands together

Use these functions with chaining commands togehter with the operator.

Details

Other functions that may be used in the chain of commands include: param, init, update, ev. or any other function that will take the output of the preceeding command as it's first argument.

See also

data_set, idata_set

Examples

mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

data(exidata)
data(exTheoph)

out <- mod %>% data_set(exTheoph) %>% mrgsim()

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

out <- mod %>% carry_out(evid) %>% ev(amt=100, cmt=1) %>% mrgsim()

#> Error in x@args[["carry.out"]] <- as_character_args(match.call()[-1]): no slot of name "args" for this object of class "standardGeneric"

out <- mod %>% Req(CP,RESP) %>% mrgsim()

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("Req"), generic = structure("Req", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'Req' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘Req’ for signature ‘"standardGeneric"’