Set the

Set the data argument for mrgsim.

data_set(x, data, ...)

# S4 method for mrgmod,data.frame
data_set(x, data, subset = TRUE,
  select = TRUE, ...)

# S4 method for mrgmod,ANY
data_set(x, data, ...)

# S4 method for mrgmod,missing
data_set(x, object, ...)

Arguments

x

model object

data

data set

...

passed along

subset

passed to dplyr::filter_; retain only certain rows in the data set

select

passed to dplyr::select_; retain only certain columns in the data set

object

character name of an object existing in $ENV to use for the data set

Details

Input data sets are R data frames that can include columns with any valid name, however columns with selected names are recognized by mrgsolve and incorporated into the simulation.

ID specifies the subject ID and is required for every input data set.

When columns have the same name as parameters ($PARAM in the model specification file), the values in those columns will be used to update the corresponding parameter as the simulation progresses.

Input data set may include the following columns related to PK dosing events: time, cmt, amt, rate, ii, addl, ss. time and cmt (and ID) are required columns in the input data set. time is the observation or event time, cmt is the compartment number (see init), amt is the dosing amount, rate is the infusion rate, ii is the dosing interval, addl specifies additional doses to administer, and ss is a flag for steady state dosing. These column names operate similarly to other non-linear mixed effects modeling software, but note that (except for ID) the column names related to PK dosing must be lower case.

Only numeric data can be brought in to the problem. Any non-numeric data columns will be dropped with warning.

See exdatasets for different example data sets.

Examples

mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve.git/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

data <- expand.ev(ID=1:3, amt=c(10,20))

mod %>% data_set(data, ID > 1) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’