Select and modify a data set for simulation

Select and modify a data set for simulation

data_set(x, data, ...)

# S4 method for mrgmod,data.frame
data_set(x, data, .subset = TRUE,
  .select = TRUE, object = NULL, need = NULL, ...)

# S4 method for mrgmod,ANY
data_set(x, data, ...)

# S4 method for mrgmod,ev
data_set(x, data, ...)

# S4 method for mrgmod,missing
data_set(x, object, ...)

Arguments

x
model object

data
data set

...
passed along

.subset
an unquoted expression passed to dplyr::filter; retain only certain rows in the data set

.select
passed to dplyr::select; retain only certain columns in the data set; this should be the result of a call to dplyr::vars()

object
character name of an object existing in $ENV to use for the data set

need
passed to inventory

Details

Input data sets are R data frames that can include columns with any valid name, however columns with selected names are treated specially by mrgsolve and incorporated into the simulation.

ID specifies the subject ID and is required for every input data set.

When columns have the same name as parameters ($PARAM in the model specification file), the values in those columns will be used to update the corresponding parameter as the simulation progresses.

Input data set may include the following columns related to PK dosing events: time, cmt, amt, rate, ii, addl, ss. time and cmt (and ID) are required columns in the input data set. time is the observation or event time, cmt is the compartment number (see init), amt is the dosing amount, rate is the infusion rate, ii is the dosing interval, addl specifies additional doses to administer, and ss is a flag for steady state dosing. These column names operate similarly to other non-linear mixed effects modeling software. Upper case PK dosing column names including TIME, CMT, AMT, RATE, II, ADDL, SS are also recognized. However, an error will be generated if a mix of upper case and lower case columns are found.

Only numeric data can be brought in to the problem. Any non-numeric data columns will be dropped with warning.

See exdatasets for different example data sets.

See also

Examples

mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

data <- expand.ev(ID=1:3, amt=c(10,20))

mod %>% data_set(data, ID > 1) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     new("nonstandardGenericFunction", .Data = function (x, data,         ...)     {        standardGeneric("data_set")    }, generic = structure("data_set", package = "mrgsolve"),         package = "mrgsolve", group = list(), valueClass = character(0),         signature = c("x", "data"), default = NULL, skeleton = (function (x,             data, ...)         stop("invalid call in method dispatch to 'data_set' (no default method)",             domain = NA))(x, data, ...)), <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

data(extran1)
head(extran1)

#>   ID  amt cmt time addl ii rate evid
#> 1  1 1000   1    0    3 24    0    1
#> 2  2 1000   2    0    0  0   20    1
#> 3  3 1000   1    0    0  0    0    1
#> 4  3  500   1   24    0  0    0    1
#> 5  3  500   1   48    0  0    0    1
#> 6  3 1000   1   72    0  0    0    1

mod %>% data_set(extran1) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     new("nonstandardGenericFunction", .Data = function (x, data,         ...)     {        standardGeneric("data_set")    }, generic = structure("data_set", package = "mrgsolve"),         package = "mrgsolve", group = list(), valueClass = character(0),         signature = c("x", "data"), default = NULL, skeleton = (function (x,             data, ...)         stop("invalid call in method dispatch to 'data_set' (no default method)",             domain = NA))(x, data, ...)), <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

mod %>% mrgsim(data=extran1)

#> Error in mrgsim(., data = extran1): no slot of name "args" for this object of class "standardGeneric"