Event objects for simulating PK and other interventions

An event object contains one or more dosing records that makes up a dosing regimen. The event object may include an ID column, but it usually doesn't. Event objects have operators and other functions that can take simple or elementary event sequences and combine them in potentially more complicated ways.

ev(x, ...)

# S4 method for mrgmod
ev(x, object = NULL, ...)

# S4 method for missing
ev(time = 0, amt, evid = 1, cmt = 1,
  ID = numeric(0), replicate = TRUE, until = NULL, realize_addl = FALSE,
  ...)

# S4 method for ev
ev(x, realize_addl = FALSE, ...)

Arguments

x
a model object

...
passed on

object
passed to show

time
event time

amt
dose amount

evid
event ID

cmt
compartment

ID
subject ID

replicate
logical; if TRUE, events will be replicated for each individual in ID

until
the expected maximum observation time for this regimen

realize_addl
if FALSE (default), no change to addl doses. If TRUE, addl doses are made explicit with realize_addl

Value

events object

Details

Required items in events objects include time, amt, evid and cmt.

If not supplied, evid is assumed to be 1.

If not supplied, cmt is assumed to be 1.

If not supplied, time is assumed to be 0.

If amt is not supplied, an error will be generated.

Other items can include ii, ss, and addl (see data_set for details on all of these items).

ID may be specified as a vector.

If replicate is TRUE (default), thenthe events regimen is replicated for each ID; otherwise, the number of event rows must match the number of IDs entered

See also

Examples

mod <- mrgsolve:::house()
mod <- mod %>% ev(amt=1000, time=0, cmt=1)
events(mod)

#> Warning: events are no longer included in the model object

loading <- ev(time=0, cmt=1, amt=1000)
maint <- ev(time=12, cmt=1, amt=500, ii=12, addl=10)
loading + maint

#> Events:
#>   time cmt  amt evid ii addl
#> 1    0   1 1000    1  0    0
#> 2   12   1  500    1 12   10

ev(ID=1:10, cmt=1, time=0, amt=100)

#> Events:
#>    ID time cmt amt evid
#> 1   1    0   1 100    1
#> 2   2    0   1 100    1
#> 3   3    0   1 100    1
#> 4   4    0   1 100    1
#> 5   5    0   1 100    1
#> 6   6    0   1 100    1
#> 7   7    0   1 100    1
#> 8   8    0   1 100    1
#> 9   9    0   1 100    1
#> 10 10    0   1 100    1