Event objects for simulating PK and other interventions.

Events can either be specified when the model object is created (with mrgmod) or by updating an existing model object (with update).

events(x, ...)

ev(x, ...)

as.ev(x, ...)

# S4 method for ev
names(x)

# S3 method for ev
mutate(.data, ...)

nrow.ev(x)

# S4 method for mrgmod
events(x, ...)

# S4 method for mrgmod
ev(x, object = NULL, ...)

# S4 method for missing
ev(time = 0, evid = 1, ID = numeric(0), cmt = 1,
  replicate = TRUE, until = NULL, realize_addl = FALSE, ...)

# S4 method for ev
ev(x, realize_addl = FALSE, ...)

# S4 method for data.frame
as.ev(x, keep_id = TRUE, clean = FALSE, ...)

# S4 method for ev
as.ev(x, ...)

# S3 method for ev
as.matrix(x, ...)

# S3 method for ev
as.data.frame(x, row.names = NULL, optional = FALSE,
  add_ID = NULL, ...)

# S4 method for ev
show(object)

Arguments

x
mrgmodel object

...
passed on

.data
the event object

object
passed to show

time
event time

evid
event ID

ID
subject ID

cmt
compartment

replicate
logical; if TRUE, events will be replicated for each individual in ID

until
the expected maximum observation time for this regimen

realize_addl
if FALSE (default), no change to addl doses. If TRUE, addl doses are made explicit with realize_addl

keep_id
if TRUE, ID column is retained if it exists

clean
if TRUE, only dosing or ID information is retained in the result

row.names
passed to as.data.frame

optional
passed to as.data.frame

add_ID
numeric ID of length 1 used to add ID column only if one doesn't already exist

Value

Returns a user-defined data frame of events that should be suitable for passing into lsoda. If events are stored as a data frame, events returns the data frame. If events are stored as a function that generates the data frame, events calls the function and passes return back to the user.

events object

Details

Required input for creating events objects include time and cmt

If not supplied, evid is assumed to be 1.

If not supplied, cmt is assumed to be 1.

If not supplied, time is assumed to be 0.

ID may be specified as a vector.

If replicate is TRUE (default), thenthe events regimen is replicated for each ID; otherwise, the number of event rows must match the number of IDs entered

Examples

mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

mod <- mod %>% ev(amt=1000, time=0, cmt=1)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

events(mod)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("events"), generic = structure("events", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'events' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘events’ for signature ‘"standardGeneric"’

loading <- ev(time=0, cmt=1, amt=1000)
maint <- ev(time=12, cmt=1, amt=500, ii=12, addl=10)
loading + maint

#> Events:
#>   time cmt  amt evid ii addl
#> 1    0   1 1000    1  0    0
#> 2   12   1  500    1 12   10

ev(ID=1:10, cmt=1, time=0, amt=100)

#> Events:
#>    ID time cmt amt evid
#> 1   1    0   1 100    1
#> 2   2    0   1 100    1
#> 3   3    0   1 100    1
#> 4   4    0   1 100    1
#> 5   5    0   1 100    1
#> 6   6    0   1 100    1
#> 7   7    0   1 100    1
#> 8   8    0   1 100    1
#> 9   9    0   1 100    1
#> 10 10    0   1 100    1