Example input data sets.

Example input data sets.

data(exidata)

data(extran1)

data(extran2)

data(extran3)

data(exTheoph)

data(exBoot)

Details

exidata holds individual-level parameters and other data items, one per row

extran1 is a "condensed" data set

extran2 is a full dataset

extran3 is a full dataset with parameters

exTheoph is the theophylline data set, ready for input into mrgsolve

exBoot a set of bootstrap parameter estimates

Examples

mod <- mrgsolve:::house() %>% update(end=240) %>% Req(CP)

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve.git/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

## Full data set
data(exTheoph)
out <- mod %>% data_set(exTheoph) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

plot(out)

#> Error in plot(out): object 'out' not found

## Condensed: mrgsolve fills in the observations
data(extran1)
out <- mod %>% data_set(extran1) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

plot(out)

#> Error in plot(out): object 'out' not found

## Add a parameter to the data set
stopifnot(require(dplyr))
data <- extran1 %>% distinct(ID) %>% select(ID) %>%
  mutate(CL=exp(log(1.5) + rnorm(nrow(.), 0,sqrt(0.1)))) %>%
  left_join(extran1,.)

#> Joining, by = "ID"

data

#>   ID  amt cmt time addl ii rate evid        CL
#> 1  1 1000   1    0    3 24    0    1 0.9634184
#> 2  2 1000   2    0    0  0   20    1 1.6261308
#> 3  3 1000   1    0    0  0    0    1 0.6940125
#> 4  3  500   1   24    0  0    0    1 0.6940125
#> 5  3  500   1   48    0  0    0    1 0.6940125
#> 6  3 1000   1   72    0  0    0    1 0.6940125
#> 7  4 2000   2    0    2 48  100    1 1.4973596
#> 8  5 1000   1    0    0  0    0    1 1.8257983
#> 9  5 5000   1   24    0  0   60    1 1.8257983

out <- mod %>% data_set(data) %>% carry.out(CL) %>%  mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

plot(out)

#> Error in plot(out): object 'out' not found

## idata
data(exidata)
out <- mod %>% idata_set(exidata) %>% ev(amt=100,ii=24,addl=10) %>% mrgsim

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("idata_set"), generic = structure("idata_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'idata_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘idata_set’ for signature ‘"standardGeneric", "data.frame"’

plot(out, CP~time|ID)

#> Error in plot(out, CP ~ time | ID): object 'out' not found