Write, compile, and load model code.

This is a convenience function that ultimately calls mread.

mcode_cache(model, code, project = tempdir(), ...)

mcode(model, code, project = tempdir(), ...)

Arguments

model

model name

code

character string specifying a mrgsolve model

project

project name

...

passed to mread

Details

Note that the arguments are in slightly different order than mread. The default project is tempdir().

Examples

code <- '
$CMT DEPOT CENT
$PKMODEL ncmt=1, depot=TRUE
$MAIN
double CL = 1;
double V = 20;
double KA = 1;
'

mod <- mcode("example",code)