modlib: Pharmacokinetic models

modlib: Pharmacokinetic models

Arguments

...

passed to update

Value

an object of class packmod

Details

See modlib_details for more detailed descriptions of parameters and compartments.

The pk1cmt model is parameterized in terms of CL, VC, KA1 and KA2 and uses compartments EV1, EV2, and CENT. The pk2cmt model adds a PERIPH compartment and parameters Q and VP to that of the one-compartment model. Likewise, the three-compartment model (pk3cmt) adds PERIPH2 and parameters Q2 and VP2 to that of the two-compartment models. All pk models also have parameters VMAX (defaulting to zero, no non-linear clearance) and KM.

Model description

All pk models have two extravascular dosing compartments and potential for linear and nonlinear clearance.

pk1cmt: one compartment pk model using ODEs

pk2cmt: two compartment pk model using ODEs

pk3cmt: three compartment pk model using ODEs

pk1: one compartment pk model in closed-form

pk2: two compartment pk model in closed-form

popex: a simple population pk model