Simulate from a model object

This function sets up the simulation run from data stored in the model object as well as arguments passed in. Note that there are several non-formal arguments to this function that can be used to customize the simulation run and it's output. Use mrgsim_df to return a data frame rather than mrgsims object.

mrgsim(x, data = NULL, idata = NULL, events = NULL, nid = 1, ...)

mrgsim_df(...)

Arguments

x
the model objects

data
NMTRAN-like data set (see data_set)

idata
a matrix or data frame of model parameters, one parameter per row (see idata_set)

events
an event object

nid
integer number of individuals to simulate; only used if idata and data are missing

...
passed to update

Value

an object of class mrgsims

Details

Both data and idata will be coreced to numeric matrix

carry.out can be used to insert data columns into the output data set. This is partially dependent on the nature of the data brought into the problem.

When using data and idata together, an error is generated if an ID occurs in data but not idata. Also, when looking up data in idata, ID in idata is assumed to be uniquely keyed to ID in data. No error is generated if ID is duplicated in data; parameters will be used from the first occurrence found in idata.

carry.out: idata is assumed to be individual-level and variables that are carried from idata are repeated throughout the invidivual's simulated data. Variables carried from data are carried via last-observation carry forward. NA is returned from observations that are inserted into simulated output that occur prior to the first record in data.

Additional arguments

mtime numeric vector of times where the model is evaluated (with solver reset), but results are not included in simulated output

Request a vector of compartment or table names to take in simulated output; if this is specified, request is ignored

obsonly omit records with evid != 0 from simulated output

obsaug logical; when TRUE and a full data set is used, the simulated output is augmented with an observation at each time in stime(). When using obsaug, a flag indicating augmented observations can be requested by including a.u.g in carry.out

recsort Default value is 1. Possible values are 1,2,3,4: 1 and 2 put doses in a data set after padded observations at the same time; 3 and 4 put those doses before padded observations at the same time. 2 and 4 will put doses scheduled through addl after observations at the same time; 1 and 3 put doses scheduled through addl before observations at the same time. recsort will not change the order of your input data set if both doses and observations are given.

filbak For each ID, carry the first record data backward to start of the simulation

tad logical; when TRUE a column is added to simulated output is added showing the time since the last dose. Only data records with evid == 1 will be considered doses for the purposes of tad calculation.

nocb if TRUE (default), time-varying items in a data set will be implemented as next observation carried back; if FALSE time-varying items in a data set will be implemented as last observation carried forward.

See also

mrgsim_variants

Examples

## example("mrgsim")

e <- ev(amt = 1000)
mod <- mrgsolve:::house()

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

out <- mod %>% ev(e) %>% mrgsim()

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out)

#> Error in plot(out): object 'out' not found

out <- mod %>% ev(e) %>% mrgsim(end=22)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

data(exTheoph)

out <- mod %>% data_set(exTheoph) %>% mrgsim()

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- mod %>% mrgsim(data=exTheoph)

#> Error in mrgsim(., data = exTheoph): no slot of name "args" for this object of class "standardGeneric"

out <- mrgsim(mod, data=exTheoph, obsonly=TRUE)

#> Error in mrgsim(mod, data = exTheoph, obsonly = TRUE): no slot of name "args" for this object of class "standardGeneric"

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- mod %>% mrgsim(data=exTheoph, obsaug=TRUE, carry.out="a.u.g")

#> Error in mrgsim(., data = exTheoph, obsaug = TRUE, carry.out = "a.u.g"): no slot of name "args" for this object of class "standardGeneric"

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- mod %>% ev(e) %>% mrgsim(req="CENT")

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

out <- mrgsim(mod, Req="CP,RESP", events = e)

#> Error in mrgsim(mod, Req = "CP,RESP", events = e): no slot of name "args" for this object of class "standardGeneric"

out

#> Error in eval(expr, envir, enclos): object 'out' not found