mrgsolve

mrgsolve is an R package maintained under the auspices of Metrum Research Group, LLC, that facilitates simulation from models based on systems of ordinary differential equations (ODE) that are typically employed for understanding pharmacokinetics, pharmacodynamics, and systems biology and pharmacology. mrgsovle consists of computer code written in the R and C++ languages, providing an interface to the DLSODA differential equation solver (written in FORTRAN) provided through ODEPACK - A Systematized Collection of ODE Solvers.

Input data sets

See data_set for help creating input data sets. See exdatasets for example input data sets.

Package help

Package index, including a listing of all functions

Reserved words in mrgsolve: reserved

About the model object

The model object has class mrgmod.

Handling simulated output

See mrgsims for methods to use with simulated output.

About the solver used by <code>mrgsolve</code>

See: aboutsolver

Examples

## example("mrgsolve")

mod <- mrgsolve:::house(delta=0.1)  %>% param(CL=0.5)

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

events <-  ev(amt=1000, cmt=1, addl=5, ii=24)

events

#> Events:
#>   time cmt  amt addl ii evid
#> 1    0   1 1000    5 24    1

mod

#> standardGeneric for "mod" defined from package "mrgsolve"
#> 
#> function (x, ...) 
#> standardGeneric("mod")
#> <environment: 0x7fcc9996d718>
#> Methods may be defined for arguments: x
#> Use  showMethods("mod")  for currently available ones.

see(mod)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("see"), generic = structure("see", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'see' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘see’ for signature ‘"standardGeneric"’

stime(mod)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("stime"), generic = structure("stime", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'stime' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘stime’ for signature ‘"standardGeneric"’

param(mod)

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “parameter_list” object: 1: invalid names
#> invalid class “parameter_list” object: 2: all parameters require names
#> invalid class “parameter_list” object: 3: all parameters must be single numbers

init(mod)

#> Problem with names:

#> x,...,

#> Parameter values without names:

#> [[1]]
#> standardGeneric("mod")
#> 

#> Error in validObject(.Object): invalid class “cmt_list” object: 1: invalid names
#> invalid class “cmt_list” object: 2: all parameters require names
#> invalid class “cmt_list” object: 3: all parameters must be single numbers

out <- mod %>% ev(events) %>% mrgsim(end=168)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

out

#> Error in eval(expr, envir, enclos): object 'out' not found

head(out)

#> Error in head(out): object 'out' not found

tail(out)

#> Error in tail(out): object 'out' not found

dim(out)

#> Error in eval(expr, envir, enclos): object 'out' not found

plot(out, GUT+CP~.)

#> Error in plot(out, GUT + CP ~ .): object 'out' not found

sims <- as.data.frame(out)

#> Error in as.data.frame(out): object 'out' not found

t72 <- subset(sims, time==72)

#> Error in subset(sims, time == 72): object 'sims' not found

str(t72)

#> Error in str(t72): object 't72' not found

idata <- data.frame(ID=c(1,2,3), CL=c(0.5,1,2),VC=12)
out <- mod %>% ev(events) %>% mrgsim(end=168, idata=idata, req="")

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out)

#> Error in plot(out): object 'out' not found

out <- mod %>% ev(events) %>% mrgsim(carry.out="amt,evid,cmt,CL")

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

head(out)

#> Error in head(out): object 'out' not found

out <-
  mod %>%
  ev() %>%
  knobs(CL=c(0.5, 1,2), amt=c(100,300,1000), cmt=1,end=48)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out, CP~., scales="same")

#> Error in plot(out, CP ~ ., scales = "same"): object 'out' not found

plot(out, RESP+CP~time|amt,groups=CL)

#> Error in plot(out, RESP + CP ~ time | amt, groups = CL): object 'out' not found

ev1 <- ev(amt=500, cmt=2,rate=10)
ev2 <- ev(amt=100, cmt=1, time=54, ii=8, addl=10)
events <- ev1+ev2
events

#> Events:
#>   time cmt amt rate evid ii addl
#> 1    0   2 500   10    1  0    0
#> 2   54   1 100    0    1  8   10

out <- mod %>% ev(ev1+ev2) %>% mrgsim(end=180, req="")

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out)

#> Error in plot(out): object 'out' not found

## "Condensed" data set
data(extran1)
extran1

#>   ID  amt cmt time addl ii rate evid
#> 1  1 1000   1    0    3 24    0    1
#> 2  2 1000   2    0    0  0   20    1
#> 3  3 1000   1    0    0  0    0    1
#> 4  3  500   1   24    0  0    0    1
#> 5  3  500   1   48    0  0    0    1
#> 6  3 1000   1   72    0  0    0    1
#> 7  4 2000   2    0    2 48  100    1
#> 8  5 1000   1    0    0  0    0    1
#> 9  5 5000   1   24    0  0   60    1

out <- mod %>% data_set(extran1) %>% mrgsim(end=200)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods"), "data.frame"),     structure(function (x, data, ...)     standardGeneric("data_set"), generic = structure("data_set", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = c("x",     "data"), default = `\001NULL\001`, skeleton = (function (x,         data, ...)     stop("invalid call in method dispatch to 'data_set' (no default method)",         domain = NA))(x, data, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘data_set’ for signature ‘"standardGeneric", "data.frame"’

plot(out,CP~time|factor(ID))

#> Error in plot(out, CP ~ time | factor(ID)): object 'out' not found

## idata
data(exidata)
exidata

#>    ID    CL    VC     KA  KOUT  IC50 FOO
#> 1   1 1.050 47.80 0.8390 2.450 1.280   4
#> 2   2 0.730 30.10 0.0684 2.510 1.840   6
#> 3   3 2.820 23.80 0.1180 3.880 2.480   5
#> 4   4 0.552 26.30 0.4950 1.180 0.977   2
#> 5   5 0.483  4.36 0.1220 2.350 0.483  10
#> 6   6 3.620 39.80 0.1260 1.890 4.240   1
#> 7   7 0.395 12.10 0.0317 1.250 0.802   8
#> 8   8 1.440 31.20 0.0931 4.030 1.310   7
#> 9   9 2.570 18.20 0.0570 0.862 1.950   3
#> 10 10 2.000  6.51 0.1540 3.220 0.699   9

out <-
  mod %>%
  ev(amt=1000, cmt=1) %>%
  idata_set(exidata) %>%
  mrgsim(end=72)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), structure(function (x,     ...) standardGeneric("ev"), generic = structure("ev", package = "mrgsolve"), package = "mrgsolve", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x,     ...) stop("invalid call in method dispatch to 'ev' (no default method)",     domain = NA))(x, ...), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘ev’ for signature ‘"standardGeneric"’

plot(out, CP~., as="log10")

#> Error in plot(out, CP ~ ., as = "log10"): object 'out' not found

# Internal model library

# NOT RUN {
mod <- mread("irm1", modlib())

mod

mod %>% ev(amt=300, ii=12, addl=3) %>% mrgsim

# }