Extract estimates from NONMEM ext file

Extract estimates from NONMEM ext file

read_nmext(run, project = getwd(), file = paste0(run, ".ext"),
  path = NULL)

Arguments

run
a run number or run identifier

project
the NONMEM project directory

file
the ext file name

path
full path and file name for ext file

Value

A list with param, omega, and sigma in a format ready to be used to update a model object.

Examples

project <- system.file("nonmem", package = "mrgsolve")

est <- read_nmext(1005, project = project)

est$param

#> $THETA1
#> [1] 9.50789
#> 
#> $THETA2
#> [1] 22.791
#> 
#> $THETA3
#> [1] 0.0714337
#> 
#> $THETA4
#> [1] 3.47451
#> 
#> $THETA5
#> [1] 113.277
#> 
#> $THETA6
#> [1] 1.02435
#> 
#> $THETA7
#> [1] 1.19212
#> 

est$omega

#>            [,1]       [,2]       [,3]
#> [1,]  0.2138790  0.1207700 -0.0116278
#> [2,]  0.1207700  0.0945105 -0.0372064
#> [3,] -0.0116278 -0.0372064  0.0465631

est$sigma

#>           [,1]     [,2]
#> [1,] 0.0491707 0.000000
#> [2,] 0.0000000 0.201769