Update the model object

After the model object is created, update various attributes.

Usage

# S4 method for mrgmod
update(object, ..., merge = TRUE, open = FALSE, data = NULL, strict = TRUE)

# S4 method for omegalist
update(object, y, ...)

# S4 method for sigmalist
update(object, y, ...)

# S4 method for parameter_list
update(object, .y, ...)

Arguments

object

a model object

...

named items to update

merge

logical indicating to merge (rather than replace) new and existing attributes

open

logical; used only when merge is TRUE and parameter list or initial conditions list is being updated; if FALSE, no new items will be added; if TRUE, the parameter list may expand.

data

a list of items to update; this list is combined with any items passed in via ...

strict

if TRUE, then an error will be generated if there is attempt to update a non-existent item

y

another object involved in update

.y

data to update

Value

The updated model object is returned.

Details

Slots that can be updated:

verbose

debug

preclean

mindt

digits

atol - absolute solver tolerance; see solversettings

rtol - relative solver tolerance; see solversettings

ss_rtol - relative tolerance when finding steady state

ss_atol - absolute tolerance when finding steady state

ixpr - see IXPR in solversettings

mxhnil - see MXHNIL in solversettings

hmin - see HMIN in solversettings

hmax - see HMAX in solversettings

maxsteps - see MXSTEP in solversettings

start, end, delta, add

tscale

request

param

init

omega

sigma

outvars

See also

Examples

if (FALSE) {
 mod <- mrgsolve::house()

 mod <- update(mod, end=120, delta=4, param=list(CL=19.1))
}